# swagger.yaml
openapi: 3.0.3
info:
  title: Trade-Back API
  description: |
    High-Performance Trading Backend API
    
    This API provides real-time market data streaming, technical indicators, 
    and historical data management for cryptocurrency trading.
    
    ## Features
    - Real-time price streaming via WebSocket (250-450+ msg/sec)
    - Historical data backfilling and management
    - Technical indicators (Enigma with Fibonacci levels)
    - Market watchlist management with auto-subscription
    - High-frequency data updates with binary protocol option
    - Session-based authentication for personalized features
    - Automatic symbol loading from Binance on startup
    
    ## WebSocket Protocol
    The WebSocket endpoint supports both JSON and binary protocols for optimal performance.
    Binary protocol reduces message size by 72% and parsing time by 16x.
    
    ## Market Watch Auto-Subscription
    When a symbol is added to the market watch list:
    1. All WebSocket connections with the same session token auto-subscribe
    2. Real-time prices start streaming immediately
    3. Only subscribed symbols receive price updates (efficient bandwidth usage)
    
    ## Rate Limits
    - REST API: 100 requests per minute per IP
    - WebSocket: No rate limit for subscriptions
    - Historical backfill: 10 requests per second (Binance limits)
  version: 1.0.0
  contact:
    name: Trading Team
    email: support@trade-back.com
  license:
    name: MIT
...
servers:
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: |
        Authenticate a user and receive a session token.
        
        **Demo Credentials:**
        - Username: `demo`
        - Password: `demo123`
        
        The session token should be included in all subsequent requests
        that require authentication using the `X-Session-Token` header.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: demo
                password:
                  type: string
                  format: password
                  example: demo123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Session token (use in X-Session-Token header)
                    example: a1b2c3d4e5f6789012345678901234567890123456789012345678901234567890
                  expires_at:
                    type: string
                    format: date-time
                    description: Token expiration time
                    example: "2024-01-16T10:30:00Z"
                  user_id:
                    type: string
                    description: User identifier
                    example: user_demo
        '400':
          description: Bad request - missing credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      description: |
        Invalidate the current session token.
        After logout, the token cannot be used for authentication.
      operationId: logout
      security:
        - SessionAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Logged out successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/session:
    get:
      tags: [Authentication]
      summary: Get session info
      description: |
        Retrieve information about the current session.
        Use this to verify if a session token is still valid.
      operationId: getSession
      security:
        - SessionAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Session information
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    example: user_demo
                  username:
                    type: string
                    example: demo
                  created_at:
                    type: string
                    format: date-time
                    example: "2024-01-15T10:30:00Z"
                  expires_at:
                    type: string
                    format: date-time
                    example: "2024-01-16T10:30:00Z"
                  valid:
                    type: boolean
                    example: true
        '401':
          description: Invalid or expired session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Invalid or expired session
...
  /symbols:
    get:
      tags: [Market Data]
      summary: Get trading symbols
      description: Retrieve a list of available trading symbols with optional filtering
      operationId: getSymbols
      parameters:
        - name: exchange
          in: query
          description: Filter by exchange
          schema:
            type: string
            example: BINANCE
        - name: type
          in: query
          description: Filter by symbol type
          schema:
            type: string
            enum: [spot, futures]
            example: spot
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: List of symbols
          content:
            application/json:
              schema:
                type: object
                properties:
                  symbols:
                    type: array
                    items:
                      $ref: '#/components/schemas/Symbol'
                  count:
                    type: integer
                    example: 350
  /symbols/{symbol}:
    get:
      tags: [Market Data]
      summary: Get symbol details
      description: Retrieve detailed information about a specific trading symbol
      operationId: getSymbol
      parameters:
        - $ref: '#/components/parameters/Symbol'
      responses:
        '200':
          description: Symbol details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Symbol'
        '404':
          $ref: '#/components/responses/NotFound'
  /symbols/{symbol}/price:
    get:
      tags: [Market Data]
      summary: Get current price
      description: Retrieve the current price and related market data for a symbol
      operationId: getPrice
      parameters:
        - $ref: '#/components/parameters/Symbol'
      responses:
        '200':
          description: Current price data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceData'
              example:
                symbol: BTCUSDT
                price: 42150.50
                bid: 42150.40
                ask: 42150.60
                volume: 15234.567
                timestamp: "2024-01-15T10:30:00Z"
                sequence: 1234567890
        '404':
          $ref: '#/components/responses/NotFound'
  /symbols/{symbol}/bars:
    get:
      tags: [Market Data]
      summary: Get historical bars
      description: Retrieve historical OHLCV (candlestick) data for a symbol
      operationId: getBars
      parameters:
        - $ref: '#/components/parameters/Symbol'
        - name: resolution
          in: query
          description: Bar resolution/timeframe
          schema:
            type: string
            enum: [1m, 5m, 15m, 30m, 1h, 4h, 1d]
            default: 1m
            example: 1h
        - name: from
          in: query
          description: Start timestamp (Unix timestamp)
          schema:
            type: integer
            format: int64
            example: 1640995200
        - name: to
          in: query
          description: End timestamp (Unix timestamp)
          schema:
            type: integer
            format: int64
            example: 1641081600
        - name: limit
          in: query
          description: Maximum number of bars to return
          schema:
            type: integer
            default: 300
            maximum: 5000
            example: 100
      responses:
        '200':
          description: Historical bars data
          content:
            application/json:
              schema:
                type: object
                properties:
                  bars:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bar'
                  symbol:
                    type: string
                    example: BTCUSDT
                  resolution:
                    type: string
                    example: 1h
  /symbols/{symbol}/enigma:
    get:
      tags: [Indicators]
      summary: Get Enigma indicator
      description: |
        Retrieve the Enigma technical indicator values including Fibonacci levels.
        The Enigma indicator calculates the current price position between ATH and ATL
        as a percentage (0-100), where:
        - 0 = Price at All-Time Low
        - 100 = Price at All-Time High
        - 50 = Price at midpoint between ATH and ATL
      operationId: getEnigma
      parameters:
        - $ref: '#/components/parameters/Symbol'
      responses:
        '200':
          description: Enigma indicator data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnigmaData'
              example:
                symbol: BTCUSDT
                level: 75.5
                ath: 69000.00
                atl: 3000.00
                current_price: 49830.00
                fib_levels:
                  "0": 3000.00
                  "23.6": 18584.80
                  "38.2": 28211.20
                  "50": 36000.00
                  "61.8": 43788.80
                  "78.6": 54874.40
                  "100": 69000.00
                timestamp: "2024-01-15T10:30:00Z"
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: Enigma calculator unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /marketwatch:
    get:
      tags: [Watchlist]
      summary: Get market watchlist
      description: |
        Retrieve the user's market watchlist with optional real-time price data.
        Requires session authentication.
        
        Use the `detailed=true` query parameter to include full market data for each symbol.
      operationId: getWatchlist
      security:
        - SessionAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionToken'
        - name: detailed
          in: query
          description: Include detailed market data for each symbol
          schema:
            type: boolean
            default: false
            example: true
      responses:
        '200':
          description: Watchlist data
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    description: Simple response (detailed=false)
                    properties:
                      symbols:
                        type: array
                        items:
                          type: string
                        example: ["BTCUSDT", "ETHUSDT", "BNBUSDT"]
                      count:
                        type: integer
                        example: 3
                  - type: object
                    description: Detailed response (detailed=true)
                    properties:
                      symbols:
                        type: array
                        items:
                          $ref: '#/components/schemas/MarketWatchSymbol'
                      count:
                        type: integer
                        example: 3
                      timestamp:
                        type: integer
                        format: int64
                        example: 1704384000
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Watchlist]
      summary: Add to watchlist
      description: |
        Add a symbol to the market watchlist. This will automatically
        subscribe all WebSocket connections with the same session token
        to receive price updates for this symbol.
      operationId: addToWatchlist
      security:
        - SessionAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbol]
              properties:
                symbol:
                  type: string
                  example: BTCUSDT
      responses:
        '200':
          description: Symbol added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  symbol:
                    type: string
                    example: BTCUSDT
        '400':
          description: Bad request or watchlist limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /marketwatch/{symbol}:
    delete:
      tags: [Watchlist]
      summary: Remove from watchlist
      description: |
        Remove a symbol from the market watchlist. This will automatically
        unsubscribe all WebSocket connections with the same session token
        from receiving price updates for this symbol.
      operationId: removeFromWatchlist
      security:
        - SessionAuth: []
      parameters:
        - $ref: '#/components/parameters/Symbol'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Symbol removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  symbol:
                    type: string
                    example: BTCUSDT
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-Token
      description: Session token for authenticated endpoints
  parameters:
    Symbol:
      name: symbol
      in: path
      required: true
      description: Trading symbol
      schema:
        type: string
        pattern: '^[A-Z]+$'
        example: BTCUSDT
    SessionToken:
      name: X-Session-Token
      in: header
      required: true
      description: Session token for authentication
      schema:
        type: string
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Session token required
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
...
    Symbol:
      type: object
      properties:
        id:
          type: integer
          example: 1
        symbol:
          type: string
          example: BTCUSDT
        exchange:
          type: string
          example: BINANCE
        full_name:
          type: string
          example: Bitcoin/Tether
        instrument_type:
          type: string
          example: SPOT
        base_currency:
          type: string
          example: BTC
        quote_currency:
          type: string
          example: USDT
        is_active:
          type: boolean
          example: true
        min_price_increment:
          type: number
          format: double
          example: 0.01
        min_quantity_increment:
          type: number
          format: double
          example: 0.00001
...